functions/requirements.txt
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.9.0
cryptography==42.0.7
deprecation==2.1.0
et-xmlfile==1.1.0
firebase-admin==6.5.0
firebase-functions==0.1.2
Flask==3.0.3
Flask-Cors==4.0.1
functions-framework==3.5.0
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-events==0.12.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==4.9.2
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.1
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
reportlab==4.2.0
requests==2.31.0
rsa==4.9
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.3
XlsxWriter==3.2.0
